--- src/test/resources/Guru/alert.feature
@@ -1,6 +1,0 @@
-Feature: Practice Alerts
-  Scenario: Click positive btn on alert
-    Given I am on Guru landing page
-    When I enter the customer id as 100
-    And I click on Submit button
-    Then I verify 'Customer Successfully Delete!' is displayed
