+++ src/test/resources/Other/Practice/jqueryFrame.feature
@@ -1,0 +1,6 @@
+Feature: Testing Frames Concept
+
+  Scenario: Locate frame on page
+    Given I am on Frames page on JqueryUI
+    When I switch to frame
+    Then I click to drag and drop box
