--- src/test/resources/spacex.feature
@@ -1,5 +1,0 @@
-Feature: space
-  Scenario: testing cucumber
-    Given I am on spacex
-    When I get title
-    Then I quit
